Adapting the SetPath function to their comment (setting LocationPath instead of URIPath)
Adapting server example to make use of the SetPath function instead of
setting the option directly. (Setting an array as options value failes
hence there is no such case in the toBytes function)

For repeating fields, it might be good to set the option in a for loop,
when receiving an array.