Return error when parsing packet with fewer bytes than TKL field indicates
This fix solves a panic which is raised when parsing a packet which
indicates a non-zero token length (TKL) but the packet itself is shorter
than the indicated token length.  In such a case, a "truncated" error is
return from UnmarshalBinary.

The bug was found with the help of go-fuzz
(https://github.com/dvyukov/go-fuzz).