Detect truncated packets when handling extended option lengths
When parsing extending option lengths, verify that enough bytes are left
in the packet.  This fixes a panic that occured when a specially crafted
truncated packet was processed.

The TestInvalidMessageParsing was refactored to allow easy additions of
such cases to the test suite.

The bug was found with the help of go-fuzz
(https://github.com/dvyukov/go-fuzz).